@@ -17,24 +17,24 @@
   "license": "MIT",
   "devDependencies": {
     "@changesets/changelog-github": "0.5.0",
-    "@changesets/cli": "2.27.3",
+    "@changesets/cli": "2.27.12",
     "@types/node": "14.18.63",
-    "@typescript-eslint/eslint-plugin": "7.11.0",
-    "@typescript-eslint/parser": "7.11.0",
+    "@typescript-eslint/eslint-plugin": "7.18.0",
+    "@typescript-eslint/parser": "7.18.0",
     "@vitest/coverage-v8": "0.34.6",
     "cpy": "10.1.0",
-    "eslint": "8.57.0",
-    "eslint-config-next": "14.2.3",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-prettier": "5.1.3",
+    "eslint": "8.57.1",
+    "eslint-config-next": "15.1.6",
+    "eslint-config-prettier": "10.0.1",
+    "eslint-plugin-prettier": "5.2.3",
     "execa": "7.2.0",
     "nx": "18.3.5",
-    "prettier": "3.2.5",
-    "prettier-plugin-tailwindcss": "0.5.14",
-    "rimraf": "5.0.7",
+    "prettier": "3.4.2",
+    "prettier-plugin-tailwindcss": "0.6.11",
+    "rimraf": "5.0.10",
     "tempy": "3.1.0",
-    "typescript": "5.4.5",
+    "typescript": "5.7.3",
     "vitest": "0.34.6"
   },
-  "packageManager": "pnpm@8.15.8"
+  "packageManager": "pnpm@8.15.9"
 }
