@@ -13,23 +13,22 @@
   "dependencies": {
     "@codemirror/lang-javascript": "6.2.2",
     "@codemirror/lang-json": "6.0.1",
-    "@codemirror/view": "6.26.3",
-    "@headlessui/react": "1.7.19",
-    "@popperjs/core": "2.11.8",
-    "@types/react": "18.3.3",
-    "@types/react-dom": "18.3.0",
-    "@uiw/react-codemirror": "4.22.1",
+    "@codemirror/view": "6.36.2",
+    "@headlessui/react": "2.2.0",
+    "@types/react": "19.0.8",
+    "@types/react-dom": "19.0.3",
+    "@uiw/react-codemirror": "4.23.7",
     "clsx": "2.1.1",
-    "next": "14.2.3",
-    "next-themes": "0.3.0",
-    "nextra": "2.13.4",
-    "nextra-theme-docs": "2.13.4",
-    "postcss": "8.4.38",
-    "postcss-lightningcss": "1.0.0",
-    "react": "18.3.1",
-    "react-docgen": "workspace:7.0.3",
-    "react-dom": "18.3.1",
-    "tailwindcss": "3.4.3"
+    "next": "15.1.6",
+    "next-themes": "0.4.4",
+    "nextra": "3.3.1",
+    "nextra-theme-docs": "3.3.1",
+    "postcss": "8.5.1",
+    "postcss-lightningcss": "1.0.1",
+    "react": "19.0.0",
+    "react-docgen": "workspace:7.1.1",
+    "react-dom": "19.0.0",
+    "tailwindcss": "3.4.17"
   },
   "browserslist": [
     "chrome 64",
