@@ -1,9 +1,11 @@
-const withNextra = require('nextra')({
+import nextra from 'nextra';
+
+const withNextra = nextra({
   theme: 'nextra-theme-docs',
   themeConfig: './src/theme.config.tsx',
 });
 
-module.exports = withNextra({
+export default withNextra({
   webpack: (config) => {
     if (!config.resolve.fallback) config.resolve.fallback = {};
 
