@@ -1,25 +1,40 @@
+import type DocumentationBuilder from './Documentation.js';
+import type {
+  BaseType,
+  Documentation,
+  ElementsType,
+  FunctionArgumentType,
+  FunctionSignatureType,
+  LiteralType,
+  MethodParameter,
+  MethodReturn,
+  ObjectSignatureType,
+  PropDescriptor,
+  PropTypeDescriptor,
+  SimpleType,
+  TSFunctionSignatureType,
+  TypeDescriptor,
+} from './Documentation.js';
+import type FileState from './FileState.js';
+import type { Config } from './config.js';
+import { createConfig, defaultHandlers } from './config.js';
+import { ERROR_CODES } from './error.js';
+import type { Handler } from './handlers/index.js';
 import * as builtinHandlers from './handlers/index.js';
-import parse from './parse.js';
-import * as builtinResolvers from './resolver/index.js';
+import type { Importer } from './importer/index.js';
 import {
   fsImporter,
   ignoreImporter,
   makeFsImporter,
 } from './importer/index.js';
-import * as utils from './utils/index.js';
-import type { Documentation } from './Documentation.js';
-import type DocumentationBuilder from './Documentation.js';
+import parse from './parse.js';
 import type {
   Resolver,
   ResolverClass,
   ResolverFunction,
 } from './resolver/index.js';
-import type { Importer } from './importer/index.js';
-import type { Handler } from './handlers/index.js';
-import type FileState from './FileState.js';
-import type { Config } from './config.js';
-import { createConfig, defaultHandlers } from './config.js';
-import { ERROR_CODES } from './error.js';
+import * as builtinResolvers from './resolver/index.js';
+import * as utils from './utils/index.js';
 
 const builtinImporters = {
   fsImporter,
@@ -66,23 +81,36 @@ export type * as babelTypes from '@babel/types';
 
 export {
   builtinHandlers,
-  builtinResolvers,
   builtinImporters,
+  builtinResolvers,
   defaultHandlers,
+  ERROR_CODES,
   makeFsImporter,
   defaultParse as parse,
   utils,
-  ERROR_CODES,
 };
 
 export type {
-  Importer,
-  Handler,
-  Resolver,
-  ResolverClass,
-  ResolverFunction,
-  FileState,
+  BaseType,
   Config,
   Documentation,
   DocumentationBuilder,
+  ElementsType,
+  FileState,
+  FunctionArgumentType,
+  FunctionSignatureType,
+  Handler,
+  Importer,
+  LiteralType,
+  MethodParameter,
+  MethodReturn,
+  ObjectSignatureType,
+  PropDescriptor,
+  PropTypeDescriptor,
+  Resolver,
+  ResolverClass,
+  ResolverFunction,
+  SimpleType,
+  TSFunctionSignatureType,
+  TypeDescriptor,
 };
