@@ -25,20 +25,28 @@ import getTypeIdentifier from './getTypeIdentifier.js';
 import isReactBuiltinReference from './isReactBuiltinReference.js';
 import unwrapBuiltinTSPropTypes from './unwrapBuiltinTSPropTypes.js';
 
-// TODO TESTME
-
 function getStatelessPropsPath(
   componentDefinition: NodePath,
 ): NodePath | undefined {
-  let value = componentDefinition;
+  if (!componentDefinition.isFunction()) return;
 
-  if (isReactForwardRefCall(value)) {
-    value = resolveToValue(value.get('arguments')[0]!);
+  return componentDefinition.get('params')[0];
+}
+
+function getForwardRefGenericsType(
+  componentDefinition: NodePath,
+): NodePath<TSType> | null {
+  const typeParameters = componentDefinition.get('typeParameters') as NodePath<
+    TSTypeParameterInstantiation | null | undefined
+  >;
+
+  if (typeParameters && typeParameters.hasNode()) {
+    const params = typeParameters.get('params');
+
+    return params[1] ?? null;
   }
 
-  if (!value.isFunction()) return;
-
-  return value.get('params')[0];
+  return null;
 }
 
 function findAssignedVariableType(
@@ -106,6 +114,19 @@ export default (componentDefinition: NodePath): NodePath[] => {
       }
     }
   } else {
+    if (isReactForwardRefCall(componentDefinition)) {
+      const genericTypeAnnotation =
+        getForwardRefGenericsType(componentDefinition);
+
+      if (genericTypeAnnotation) {
+        typePaths.push(genericTypeAnnotation);
+      }
+
+      componentDefinition = resolveToValue(
+        componentDefinition.get('arguments')[0]!,
+      );
+    }
+
     const propsParam = getStatelessPropsPath(componentDefinition);
 
     if (propsParam) {
