@@ -1,6 +1,7 @@
 import React from 'react';
 import type { DocsThemeConfig } from 'nextra-theme-docs';
 import { useRouter } from 'next/router';
+import { useConfig } from 'nextra-theme-docs';
 
 const config: DocsThemeConfig = {
   logo: <span className="text-lg font-bold">react-docgen</span>,
@@ -10,20 +11,23 @@ const config: DocsThemeConfig = {
   docsRepositoryBase:
     'https://github.com/reactjs/react-docgen/tree/main/packages/website',
   footer: {
-    text: '',
+    content: '',
   },
-  useNextSeoProps: () => {
-    const { asPath } = useRouter();
+  head() {
+    const { frontMatter, title } = useConfig();
+    const { route } = useRouter();
 
-    if (asPath !== '/') {
-      return {
-        titleTemplate: '%s \u2013 react-docgen',
-      };
-    } else {
-      return {
-        titleTemplate: 'react-docgen \u2013 React documentation generator',
-      };
-    }
+    const description =
+      frontMatter.description || 'Make beautiful websites with Next.js & MDX.';
+    const pageTitle = title + (route === '/' ? '' : ' \u2013 react-docgen');
+
+    return (
+      <>
+        <title>{pageTitle}</title>
+        <meta property="og:title" content={pageTitle} />
+        <meta property="og:description" content={description} />
+      </>
+    );
   },
 };
 
